Factor out code for displaying events
Earlier this has been implemented as part of transaction details,
but actually it should be independent. Move components to separate files
and directories.
