+++ .changelog/991.internal.md
@@ -1,0 +1,1 @@
+Refactor code for showing events
